Fix bug where page numbers in a loaded book would add spaces, update comments, remove redundant regex escapes
The new page number comments have spaces between themselves and the pagebreaks. While the pagebreaks and page number comments were being stripped, the spaces were not. This resulted in extra spaces beign inserted at the start of pages

The redundant escapes in the regex that strips unwanted characters from the author and title when generating a filename have also beeen removed.